Move apply probe to healthchecks package
Towards reducing the surface area around healthchecks
